+++ web/src/beta/components/fields/common/NumberInput/index.stories.tsx
